+++ hydra-cluster/src/HydraNode.hs
@@ -33,13 +33,14 @@ import Data.Aeson (Value (String), object, (.=))
 import qualified Data.Aeson as Aeson
 import Data.Aeson.Types (Pair)
 import qualified Data.List as List
+import Data.Text (pack)
 import qualified Data.Text as T
 import Hydra.Cluster.Faucet (FaucetLog)
 import Hydra.Cluster.Util (readConfigFile)
 import Hydra.Crypto (HydraKey)
 import Hydra.Ledger.Cardano ()
 import Hydra.Logging (Tracer, Verbosity (..), traceWith)
-import Hydra.Network (Host (Host))
+import Hydra.Network (Host (Host), NodeId (NodeId))
 import qualified Hydra.Network as Network
 import Hydra.Options (ChainConfig (..), LedgerConfig (..), RunOptions (..), defaultChainConfig, toArgs)
 import Network.HTTP.Conduit (HttpExceptionContent (ConnectionFailure), parseRequest)
@@ -283,7 +284,7 @@ withHydraNode tracer chainConfig workDir hydraNodeId hydraSKey hydraVKeys allNod
             ( hydraNodeProcess $
                 RunOptions
                   { verbosity = Verbose "HydraNode"
-                  , nodeId = fromIntegral hydraNodeId
+                  , nodeId = NodeId $ show hydraNodeId
                   , host = "127.0.0.1"
                   , port = fromIntegral $ 5000 + hydraNodeId
                   , peers
@@ -360,11 +361,7 @@ waitForNodesConnected tracer clients =
         ( \nodeId ->
             object
               [ "tag" .= String "PeerConnected"
-              , "peer"
-                  .= object
-                    [ "hostname" .= ("127.0.0.1" :: Text)
-                    , "port" .= (5000 + nodeId)
-                    ]
+              , "peer" .= String (pack $ show nodeId)
               ]
         )
         (filter (/= hydraNodeId) allNodeIds)